7. saudi-b2b-leads-validator/requirements.txt
# Data Processing
pandas
openpyxl

# Automation & Scraping
selenium
webdriver-manager

# Email Validation (Module 3)
dnspython

# Networking
requests